Extract NavMenu component and add menu E2E tests
- Move hamburger menu into NavMenu component (src/lib/components/ui/nav-menu/)
- Update home page to use <NavMenu />
- Expand faq.test.ts to cover all menu items: How it works, About, Sign out
- Sign out test uses its own auth context to avoid invalidating shared session storage
