+++ src/lib/components/ui/nav-menu/index.ts
@@ -1,0 +1,1 @@
+export { default as NavMenu } from './nav-menu.svelte';
